initalizes tool_dict to an empty dict instead of getting the full list of tools already available to the bot
If I understand the code correctly, the current version gets all the tools currently available to the bot and can only add to them, not take away. Since modify_input.agent.tools contains a list of all the tools available after the modification, the function will validate those but not register if that list does not include previously available tools. 

By initializing tool_dict to an empty dictionary, the function builds the tool list from scratch based on the available tools in modify_input.agent.tools